requirements.txt
amqp==5.0.3
asgiref==3.3.1
attrs==20.3.0
billiard==3.6.3.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
dj-database-url==0.5.0
Django==3.1.3
django-cors-headers==3.7.0
django-decouple==2.1
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
drf-spectacular==0.11.1
gunicorn==20.0.4
importlib-metadata==3.7.2
inflection==0.5.1
jsonschema==3.2.0
kombu==5.0.2
Markdown==3.3.3
Pillow==8.1.0
prompt-toolkit==3.0.14
psycopg2-binary==2.8.6
PyJWT==2.0.1
pyrsistent==0.17.3
pytz==2020.4
PyYAML==5.3.1
six==1.15.0
sqlparse==0.4.1
twitter==1.18.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.1
